MakiePublicationTheme

Documentation for MakiePublicationTheme.

MakiePublicationTheme.size_pixelgraphic

MakiePublicationTheme.size_vectorgraphic

MakiePublicationTheme.size_pixelgraphic — Method

size_pixelgraphic(size, dpi; unit=:cm)

Compute the required figure size and px_per_unit at a resolution of dpi dots-per-inch and for a given physical size in :cm or :inch. The (physical) font size will be the same as set in the theme, i.e. the default is 10 pt.

Example:

using MakiePublicationTheme
using CairoMakie

size_cm = (12.0, 8.0)
dpi = 600
size, px_per_unit = size_pixelgraphic(size_cm, dpi)

f = Figure(; size)
# lots of plotting here
save("my_12x8cm_figure.png", f; px_per_unit)

More information about matching figure and font sizes can be found in the Makie documentation.

MakiePublicationTheme.size_vectorgraphic — Method

size_vectorgraphic(size; unit=:cm)

Compute the required figure size and pt_per_unit for a given physical size in :cm or :inch. The (physical) font size will be the same as set in the theme, i.e. the default is 10 pt.

Example:

using MakiePublicationTheme
using CairoMakie

size_cm = (12.0, 8.0)
size, pt_per_unit = size_vectorgraphic(size_cm)

f = Figure(; size)
# lots of plotting here
save("my_12x8cm_figure.pdf", f; pt_per_unit)

More information about matching figure and font sizes can be found in the Makie documentation.